requirements.txt
langchain
langchain-groq
langchain-community
duckduckgo-search
arxiv
wikipedia
python-dotenv
